Many changes, works in more cases now. Types in comments.
DisableMethods in order to go deep in structs.
Strip "main." if that's the package prefix, keep all others for now.
Print types in comments.
Print zero values again (it helps see what fields there are).
Switch to using gofmt binary because it got too hard to do it manually
(due to desire to print comments too). To be replaced by go/format in
go1.1.