Added a new test based on git diff.
If this experiment is successful, it should reduce manual duplication
in testing.

To run the test: go test github.com/shurcooL/go-goon

It's based on the notion that tests are supposed to make sure you get
the same results as you expect. The test prints things to be tested to
stdout and saves them to a file, which is then compared to previous
versions of the output. If it matches, the test passes.

More precisely, for all *.go files in ./tests/, it runs:

go run file.go > file.go.out
git diff --no-ext-diff -- file.go.out

If there is zero output and zero return values, the test is considered
to pass.