Experiment to reduce manual duplication, README.md is now generated from parts.
I still need to create a tool that will allow editing the file in this
format in an easy way. For now it was put together manually.

However, as test outputs change, the README.md is guarnateed to have
the latest version embedded (assuming you run its Makefile).

This is only the beginning.