Removed hex dumping of uint8 slices. Fixes #2.
I've considered fixing this in a way that keeps the hex dump
visualization. However, it seems to fall outside of go-goon's
responsibility, which is program-to-program communication. The
program-to-human layer should be built on top IMO.

E.g., something else can take a raw []byte goon dump and visualize it
as a hex dump.