Add a note that the code mirrors (at this time) that of spew.
/cc @davecgh
This may change if I later play with the bypass_alt approach and find
it works better. However, staying with go-spew's version means it is
better tested.
At the end of the day, I would like a single (exported) best solution
for bypass (if it's possible for it to cover all use cases and needs).